move tests to store
